requirements.txt
aiogram==3.7.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
asyncpg==0.29.0
pydantic==2.7.2
diffusers==0.28.0
transformers==4.41.2
accelerate==0.30.1
cachetools==5.3.3